Convert 구조체 생성해서 convert 함수들 별도의 객체로 분리 & Json 구조체 생성해서 json파싱부분은 다른 객체에서 처리
